+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Controllers/AccountController.cs
@@ -1,16 +1,18 @@
-﻿using System.Threading.Tasks;
-using AwesomeCMSCore.Modules.Admin.Extensions;
-using AwesomeCMSCore.Modules.Admin.Models.AccountViewModels;
+﻿using System;
+using System.Threading.Tasks;
+using AwesomeCMSCore.Modules.Account.Extensions;
+using AwesomeCMSCore.Modules.Account.Models.AccountViewModels;
 using AwesomeCMSCore.Modules.Email;
 using AwesomeCMSCore.Modules.Entities.Entities;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
-namespace AwesomeCMSCore.Modules.Admin.Controllers.API
+namespace AwesomeCMSCore.Modules.Account.Controllers
 {
+
     [Authorize]
-    [Route("api/[controller]/[action]")]
     public class AccountController : Controller
     {
         private readonly UserManager<User> _userManager;
@@ -27,48 +29,62 @@ namespace AwesomeCMSCore.Modules.Admin.Controllers.API
             _emailSender = emailSender;
         }
 
-        [HttpPost]
+        [TempData]
+        public string ErrorMessage { get; set; }
+
+        [HttpGet]
         [AllowAnonymous]
-        public async Task<IActionResult> Login([FromBody] LoginViewModel model)
+        public async Task<IActionResult> Login()
         {
-            var result = await _signInManager.PasswordSignInAsync(model.Username, model.Password,
-                model.RememberMe, lockoutOnFailure: false);
-
-            if (result.Succeeded)
-            {
-                return Ok();
-            }
-            if (result.IsLockedOut)
-            {
-                return Unauthorized();
-            }
-
-            return BadRequest();
+            await HttpContext.SignOutAsync(IdentityConstants.ExternalScheme);
+            return View();
         }
-       
-        [HttpPost]
+
+        [HttpGet]
         [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Register(RegisterViewModel model, string returnUrl = null)
+        public IActionResult Lockout()
+        {
+            return View();
+        }
+
+        [HttpGet]
+        [AllowAnonymous]
+        public IActionResult Register(string returnUrl = null)
         {
             ViewData["ReturnUrl"] = returnUrl;
-            if (ModelState.IsValid)
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Logout()
+        {
+            await _signInManager.SignOutAsync();
+            return RedirectToAction("Login");
+        }
+
+        [HttpGet]
+        [AllowAnonymous]
+        public async Task<IActionResult> ConfirmEmail(string userId, string code)
+        {
+            if (userId == null || code == null)
             {
-                var user = new User { UserName = model.Email, Email = model.Email };
-                var result = await _userManager.CreateAsync(user, model.Password);
-                if (result.Succeeded)
-                {
-                    var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-                    var callbackUrl = Url.EmailConfirmationLink(user.Id, code, Request.Scheme);
-                    await _emailSender.SendEmailConfirmationAsync(model.Email, callbackUrl);
-
-                    await _signInManager.SignInAsync(user, isPersistent: false);
-                    return Ok(returnUrl);
-                }
+                return RedirectToAction("Login");
             }
+            var user = await _userManager.FindByIdAsync(userId);
+            if (user == null)
+            {
+                throw new ApplicationException($"Unable to load user with ID '{userId}'.");
+            }
+            var result = await _userManager.ConfirmEmailAsync(user, code);
+            return View(result.Succeeded ? "ConfirmEmail" : "Error");
+        }
 
-            // If we got this far, something failed, redisplay form
-            return View(model);
+        [HttpGet]
+        [AllowAnonymous]
+        public IActionResult ForgotPassword()
+        {
+            return View();
         }
 
         [HttpPost]
@@ -82,7 +98,7 @@ namespace AwesomeCMSCore.Modules.Admin.Controllers.API
                 if (user == null || !(await _userManager.IsEmailConfirmedAsync(user)))
                 {
                     // Don't reveal that the user does not exist or is not confirmed
-                    return NotFound();
+                    return RedirectToAction(nameof(ForgotPasswordConfirmation));
                 }
 
                 // For more information on how to enable account confirmation and password reset please
@@ -91,10 +107,29 @@ namespace AwesomeCMSCore.Modules.Admin.Controllers.API
                 var callbackUrl = Url.ResetPasswordCallbackLink(user.Id, code, Request.Scheme);
                 await _emailSender.SendEmailAsync(model.Email, "Reset Password",
                    $"Please reset your password by clicking here: <a href='{callbackUrl}'>link</a>");
-                return Ok();
+                return RedirectToAction(nameof(ForgotPasswordConfirmation));
             }
 
             // If we got this far, something failed, redisplay form
+            return View(model);
+        }
+
+        [HttpGet]
+        [AllowAnonymous]
+        public IActionResult ForgotPasswordConfirmation()
+        {
+            return View();
+        }
+
+        [HttpGet]
+        [AllowAnonymous]
+        public IActionResult ResetPassword(string code = null)
+        {
+            if (code == null)
+            {
+                throw new ApplicationException("A code must be supplied for password reset.");
+            }
+            var model = new ResetPasswordViewModel { Code = code };
             return View(model);
         }
 
@@ -105,25 +140,32 @@ namespace AwesomeCMSCore.Modules.Admin.Controllers.API
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest(model);
+                return View(model);
             }
             var user = await _userManager.FindByEmailAsync(model.Email);
             if (user == null)
             {
                 // Don't reveal that the user does not exist
-                return Ok();
+                return RedirectToAction(nameof(ResetPasswordConfirmation));
             }
             var result = await _userManager.ResetPasswordAsync(user, model.Code, model.Password);
             if (result.Succeeded)
             {
-                return Ok();
+                return RedirectToAction(nameof(ResetPasswordConfirmation));
             }
-            return Ok();
+
+            return View();
         }
 
         [HttpGet]
         [AllowAnonymous]
         public IActionResult ResetPasswordConfirmation()
+        {
+            return View();
+        }
+
+        [HttpGet]
+        public IActionResult AccessDenied()
         {
             return View();
         }
