+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Models/AccountViewModels/ForgotPasswordViewModel.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace AwesomeCMSCore.Modules.Account.Models.AccountViewModels
+{
+    public class ForgotPasswordViewModel
+    {
+        [Required]
+        [EmailAddress]
+        public string Email { get; set; }
+    }
+}
