+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Models/AccountViewModels/LoginViewModel.cs
@@ -1,0 +1,17 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace AwesomeCMSCore.Modules.Account.Models.AccountViewModels
+{
+    public class LoginViewModel
+    {
+        [Required]
+        public string Username { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        [Display(Name = "Remember me?")]
+        public bool RememberMe { get; set; }
+    }
+}
