+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,5 @@
-﻿@model ForgotPasswordViewModel
+﻿@using AwesomeCMSCore.Modules.Account.Models.AccountViewModels
+@model AwesomeCMSCore.Modules.Account.Models.AccountViewModels.ForgotPasswordViewModel
 @{
     ViewData["Title"] = "Forgot your password?";
 }
