+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Views/Account/Login.cshtml
@@ -1,0 +1,9 @@
+﻿@using AwesomeCMSCore.Modules.Account.Models.AccountViewModels
+@model AwesomeCMSCore.Modules.Account.Models.AccountViewModels.LoginViewModel
+@{
+    Layout = "_Layout";
+    ViewData["Title"] = "Log in";
+}
+
+<div id="loginForm">
+</div>
