+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,5 @@
-﻿@model ResetPasswordViewModel
+﻿@using AwesomeCMSCore.Modules.Account.Models.AccountViewModels
+@model AwesomeCMSCore.Modules.Account.Models.AccountViewModels.ResetPasswordViewModel
 @{
     ViewData["Title"] = "Reset password";
 }
