+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Views/_ViewImports.cshtml
@@ -1,0 +1,2 @@
+﻿@using AwesomeCMSCore.Modules.Admin
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
