+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Views/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+﻿@{
+    Layout = "_AdminLayout";
+}
