+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Admin/AwesomeCMSCore.Modules.Admin.csproj
@@ -4,23 +4,21 @@
     <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Compile Remove="Repositories\**" />
+    <EmbeddedResource Remove="Repositories\**" />
+    <None Remove="Repositories\**" />
+  </ItemGroup>
+
     <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="2.0.1" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.0.2" />
-        <PackageReference Include="AspNet.Security.OAuth.Validation" Version="2.0.0-*" />
-        <PackageReference Include="OpenIddict" Version="2.0.0-*" />
-        <PackageReference Include="OpenIddict.EntityFrameworkCore" Version="2.0.0-*" />
-        <PackageReference Include="OpenIddict.Mvc" Version="2.0.0-*" />
     </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\AwesomeCMSCore.Modules.Entities\AwesomeCMSCore.Modules.Entities.csproj" />
       <ProjectReference Include="..\AwesomeCMSCore.Modules.Helper\AwesomeCMSCore.Modules.Helper.csproj" />
       <ProjectReference Include="..\AwesomeCMSCore.Modules.Repositories\AwesomeCMSCore.Modules.Repositories.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Repositories\" />
     </ItemGroup>
 
 </Project>
