--- src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Admin/Views/Account/Login.cshtml
@@ -1,8 +1,0 @@
-﻿@model LoginViewModel
-@{
-    Layout = "_Layout";
-    ViewData["Title"] = "Log in";
-}
-
-<div id="loginForm">
-</div>
